+++ src/services/Assets.php
@@ -650,7 +650,7 @@ class Assets extends Component
      */
     public function getImagePreviewUrl(Asset $asset, int $maxWidth, int $maxHeight): string
     {
-        $isWebSafe = in_array($asset->getExtension(), Image::webSafeFormats(), true);
+        $isWebSafe = Image::isWebSafe($asset->getExtension());
         $originalWidth = (int)$asset->getWidth();
         $originalHeight = (int)$asset->getHeight();
         [$width, $height] = AssetsHelper::scaledDimensions((int)$asset->getWidth(), (int)$asset->getHeight(), $maxWidth, $maxHeight);